+++ src/Service/Ecr/src/Input/GetAuthorizationTokenRequest.php
@@ -34,10 +34,14 @@ final class GetAuthorizationTokenRequest extends Input
     }
 
     /**
+     * @deprecated
+     *
      * @return string[]
      */
     public function getRegistryIds(): array
     {
+        @trigger_error(\sprintf('The property "registryIds" of "%s" is deprecated by AWS.', __CLASS__), \E_USER_DEPRECATED);
+
         return $this->registryIds ?? [];
     }
 
@@ -67,10 +71,13 @@ final class GetAuthorizationTokenRequest extends Input
     }
 
     /**
+     * @deprecated
+     *
      * @param string[] $value
      */
     public function setRegistryIds(array $value): self
     {
+        @trigger_error(\sprintf('The property "registryIds" of "%s" is deprecated by AWS.', __CLASS__), \E_USER_DEPRECATED);
         $this->registryIds = $value;
 
         return $this;
@@ -80,6 +87,7 @@ final class GetAuthorizationTokenRequest extends Input
     {
         $payload = [];
         if (null !== $v = $this->registryIds) {
+            @trigger_error(\sprintf('The property "registryIds" of "%s" is deprecated by AWS.', __CLASS__), \E_USER_DEPRECATED);
             $index = -1;
             $payload['registryIds'] = [];
             foreach ($v as $listValue) {