+++ src/Service/Ecr/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - AWS api-change: Added us-isob-east-1 region
 
+### Deprecated
+
+- AWS api-change: `GetAuthorizationTokenRequest::$registryIds` has been deprecated
+
 ## 0.1.0
 
 First version